refactor: file 엔티티를 form 엔티티로 수정
