+++ src/entities/form/assets/icon-camera.png
